ajusting to mysql
