+++ common/models/appuser.js
@@ -3,6 +3,6 @@
 // This file is licensed under the Artistic License 2.0.
 // License text available at https://opensource.org/licenses/Artistic-2.0
 
-module.exports = function(User) {
+module.exports = function(Appuser) {
 
 };
