+++ common/models/appuser.json
@@ -1,5 +1,5 @@
 {
-  "name": "user",
+  "name": "appuser",
   "base": "User",
   "idInjection": true,
   "properties": {},
